Create API, bootstrap, install barkutils, connect to redshift
Created a pipenv environment and installed needed packages. Spun up Flask API and bootstraped the initialization into an executable. Installed just the dependencies needed to use bark utils to connect to redshift. Added the a couple get and a post request functions, mostly placeholders. Defined the anticipated structure of the returned json and tested.