Presentation Skeleton, backend poc, etl examples
Added thoughts on the project proposal, functionallity, example use cases and more.
Created all of the tables for the DB and populated them with sample data to test the API/front end.
Remains to write an SP to populate daily treatments. after that, write needed api functions and build a CRA frontend for demo.